requirements.txt
Django==1.5.1
South==0.8.1
dj-database-url==0.2.2
django-tastypie==0.9.15
gunicorn==17.5
mimeparse==0.1.3
psycopg2==2.5.1
pygeocoder==1.2.1
python-dateutil==2.1
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
